Create waf_update_certificate.yaml
do not use this in enterprise environments, this is for demo purposes only to walk through how the certificate update takes place, when used, this could expose CA and key information in plaintext in the logs. again, this is for demo purposes only as an example of how AAP can communicate with a WAF -- not for enterprise use.